divisor-wave-python/requirements.txt
# Divisor Wave Analysis API - Performance Optimized Requirements
# Enhanced version with JIT compilation, GPU acceleration, and multiprocessing

# Core mathematical libraries
numpy>=2.3.4
scipy>=1.16.3
matplotlib>=3.10.7
mpmath>=1.3.0
sympy>=1.12.0           # Symbolic mathematics for LaTeX parsing
antlr4-python3-runtime>=4.11.0  # LaTeX parser backend

# Performance optimization libraries
numba>=0.56.0          # JIT compilation for critical functions
# cupy>=13.6.0           # GPU acceleration (CUDA required)
# Alternative GPU options:
cupy-cuda11x>=13.6.0  # For CUDA 11.x
# cupy-cuda12x>=13.6.0  # For CUDA 12.x
# cupy-cuda13x>=13.6.0  # For CUDA 13.x

# Web API framework
fastapi>=0.121.0
uvicorn>=0.38.0
pydantic>=2.12.3

# Additional utility libraries
python-multipart>=0.0.20
requests>=2.32.5
psutil>=7.1.3          # System monitoring

# Development and testing
pytest>=8.4.2
pytest-asyncio>=1.2.0

# Optional: Advanced optimization
# jax>=0.8.0            # Alternative to numba for XLA compilation
# jaxlib>=0.8.0         # JAX backend
# torch>=2.9.0          # PyTorch for advanced GPU operations

# Installation notes:
# 1. For NVIDIA GPU support: pip install cupy-cuda11x (or appropriate CUDA version)
# 2. For Intel/AMD GPUs: pip install intel-extension-for-pytorch (experimental)
# 3. For Apple Silicon: Use built-in acceleration via numpy/scipy
# 4. For CPU-only systems: All multiprocessing features will work automatically